trasferimento codice custom sd2, parte 1

+++ scripts/world/spell_scripts.cpp
@@ -515,12 +515,12 @@ bool EffectDummyCreature_spell_dummy_npc(Unit* pCaster, uint32 uiSpellId, SpellE
                     if (pCreatureTarget->GetEntry() == NPC_ELK)
                     {
                         pCreatureTarget->CastSpell(pCreatureTarget, SPELL_LIQUID_FIRE_AURA, true);
-                        ((Player*)pCaster)->KilledMonsterCredit(NPC_ELK_BUNNY);
+                        ((Player*)pCaster)->KilledMonsterCredit(uint32(NPC_ELK_BUNNY));
                     }
                     else if (pCreatureTarget->GetEntry() == NPC_GRIZZLY)
                     {
                         pCreatureTarget->CastSpell(pCreatureTarget, SPELL_LIQUID_FIRE_AURA, true);
-                        ((Player*)pCaster)->KilledMonsterCredit(NPC_GRIZZLY_BUNNY);
+                        ((Player*)pCaster)->KilledMonsterCredit(uint32(NPC_GRIZZLY_BUNNY));
                     }
                 }
                 return true;
